requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flask-cors==5.0.0
python-dotenv==1.0.1
